% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fun.R
\name{filter_dup}
\alias{filter_dup}
\title{Filter near-duplicate positions}
\usage{
filter_dup(data, step.time = 2/60, step.dist = 0.001)
}
\arguments{
\item{data}{data.frame with location data}

\item{step.time}{time difference to consider duplicated positions, in hours}

\item{step.dist}{spatial distance to consider duplicated poisitions, in km}
}
\value{
new data.frame with filtered data
}
\description{
Filter near-duplicate positions
}
