% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fun.R
\name{filter_speed}
\alias{filter_speed}
\title{Speed filter}
\usage{
filter_speed(data, vmax = 10.8, method = 1)
}
\arguments{
\item{data}{data.frame with location data}

\item{vmax}{value of the maximum of velocity using, in km/h}

\item{method}{see ?ddfilter.speed}
}
\value{
new data.frame with filtered data
}
\description{
Speed filter
}
