% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpeciesCode.R
\name{getSpeciesCode}
\alias{getSpeciesCode}
\title{Generate species code}
\usage{
getSpeciesCode(x)
}
\arguments{
\item{x}{character, scientific name}
}
\value{
A character string with species code.
}
\description{
Generates a 6 character species code based on Scientific name of a species.
}
