% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat2L0}
\alias{stat2L0}
\title{Proces STAT location data}
\usage{
stat2L0(stat_file, date_deploy = NULL)
}
\arguments{
\item{stat_file}{Path to csv file from STAT.}

\item{date_deploy}{POSIXct date time. If NULL, the function will not filter data by date.}
}
\value{
A data.frame with standardized names and parsed times from L0 product.
}
\description{
Proces STAT location data
}
