% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarizeTrips}
\alias{summarizeTrips}
\title{Summarize trip segment data}
\usage{
summarizeTrips(
  data,
  id = "organismID",
  trip = "tripID",
  date = "time",
  lon = "longitude",
  lat = "latitude"
)
}
\arguments{
\item{data}{Data.frame.}

\item{id}{column name of the animal id.}

\item{date}{column name of the date, with POSIXct format.}

\item{lon}{column name of longitude.}

\item{lat}{column name of latitude.}
}
\value{
A data.frame with summary data per trip.
}
\description{
Summarize trip segment data
}
