% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fun.R
\name{timedif.segment}
\alias{timedif.segment}
\title{Segment trips based on temporal threshold}
\usage{
timedif.segment(x, thrs)
}
\arguments{
\item{x}{POSIXct date time}

\item{thrs}{time threshold criteria to split into segments, in hours}
}
\value{
A vector with segments IDs
}
\description{
Segment trips based on temporal threshold
}
