% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc2L0.R
\name{wcLoc2L0}
\alias{wcLoc2L0}
\title{Proces Wildlife Computers location data}
\usage{
wcLoc2L0(data, locale = "English", date_deploy = NULL)
}
\arguments{
\item{data}{Data.frame.}

\item{locale}{System time locale.}

\item{deploymentDateTime}{POSIXct date time. If NULL, the function will not filter data by date.}
}
\value{
A data.frame with standardized names and parsed times to accomodate L0 product.
}
\description{
Proces Wildlife Computers location data
}
